+++ src/main/java/com/neuedu/demo/dao/LessonMapper.java
@@ -6,7 +6,7 @@ import com.neuedu.demo.domain.LessonInfo;
 
 @Mapper
 public interface LessonMapper {
-	public int insertLesson(LessonInfo lessoninfo);
+	public long insertLesson(LessonInfo lessoninfo);
 	public void deleteLesson(Long id);
 	public void updateLesson(LessonInfo lessoninfo);
 	public LessonInfo queryLessonById(Long id);
